Implement DeviceService and Test

+++ src/main/java/project/univAlarm/service/UserService.java
@@ -1,6 +1,7 @@
 package project.univAlarm.service;
 
 import lombok.RequiredArgsConstructor;
+import org.springframework.security.access.AccessDeniedException;
 import org.springframework.stereotype.Service;
 import project.univAlarm.domain.User;
 import project.univAlarm.repository.UserRepository;
@@ -15,7 +16,10 @@ public class UserService {
                 .orElseThrow(()->new IllegalArgumentException("해당 사용자가 존재하지 않습니다. id="+id));
     }
 
-    public void delete(Long id){
+    public void delete(User user,Long id){
+        if(!user.getId().equals(id)){
+            throw new AccessDeniedException("Access denied");
+        }
         userRepository.deleteById(id);
     }
 }
